requirements.txt
Flask==2.3.3
python-docx==0.8.11
pdfplumber==0.9.0
spacy==3.7.2
scikit-learn==1.3.0
sentence-transformers==2.2.2
transformers==4.21.0
torch==2.0.1
huggingface_hub==0.16.4
pandas==2.0.3
numpy==1.24.3
werkzeug==2.3.7
python-multipart==0.0.6
gunicorn==21.2.0